Cast boolean fields to integer and standardize casing in SQL queries for consistency and improved readability. Query 141 to 185

+++ postgres_sql_queries/query_162.sql
@@ -19,7 +19,7 @@ SELECT
 			,scriptweight
 			,cutoffscore
 			,assessmentphase_id
-			,issilentscore
+			,issilentscore::integer
 			,<BATCHDATE> AS ExtractDate
 			,<BATCHDATE> AS BatchDate
 			,<SOURCESKEY> AS SourceSkey
